Improved paddleLong and paddleShort looks
